Implement inplace writes for DSR backend.
Summary:
This improves the performance of DSR packet writing. The thread local contiguous buffer is transparently allocated on the first write.

We store the relevant data in a fake connection state so as to not change the existing (somewhat janky) batch writer interface. In the future we should refactor this so it's less ugly.

Reviewed By: kvtsoy

Differential Revision: D46765252

fbshipit-source-id: cf0bb66a3c4ea7b76ade4e99cb65ce42774ed810