Add BBRv2 Congestion Controller
Summary:
This implements the BBRv2 congestion controller for Mvfst. This implementation aims at staying as close as possible to the naming and organization used in the latest draft. Wherever there are gaps in the spec or the pseudocode, this attempts to fill them with reasonable assumptions.

https://datatracker.ietf.org/doc/html/draft-cardwell-iccrg-bbr-congestion-control-02

Besides the spec differences, this implementation differs from Mvfst's BBRv1 as follows. These differences may be revisited in the future:
- The pacing rate is exclusively controlled by the CCA. No further scaling of the pacer rate is enforced. This enables BBRv2 control loop to keep tight control over queueing delay.
- BBRv2 does not rely on an external bandwidth or RTT sampler. Mvfst's BBRv1 decouples these which makes tracking the state a bit harder.

Reviewed By: mjoras

Differential Revision: D44599226

fbshipit-source-id: 1e488bd936db5826b73ff6205f988dd40e908eba