Fix BBRv2 ProbeBW cycle counting
Summary:
The draft description mentions the max bw filter using the probe bw cycle count, but the pseudo code was using the round count. This change makes it match the description rather than the pseudo code. This fixes the situation where the pacing margin leads to a downward spiral in the target pacing rate (lower pacing rate -> lower bw estimate -> lower pacing rate -> etc.).

Also, the ack phase is not used so removing it.

Reviewed By: mjoras

Differential Revision: D46974994

fbshipit-source-id: 7936183e5cd10a75f88ede33045d7f4461fc27e0