QuicCubic track when cwnd blocked and add new config to not grow cwnd when not blocked
Summary:
Currently, QuicCubic grows its congestion window whenever:
- Packets are acked in Hystart.
- The connection is not idle in Congestion Avoidance.

QuicCubic does not track application limited events.

This leads to a situation where the congestion window continues to grows with incoming acks even through the connection never actually utilized the full congestion window. Consequently, if the connection has enough data later on, it can send big bursts using a large congestion window that has not been verified, which can lead to excessive losses.

Cwnd vs Bytes in flight without new config:
https://pxl.cl/2VwqC

This change enables QuicCubic to track when the connection send rate is blocked by the congestion window and adds a new configuration where the congestion window grows only when the connection is cwnd blocked. This means that at any point in time, the congestion window accurately reflects an amount of outstanding bytes that has been tested on the connection.

Cwnd vs. Bytes in flight with new config:
https://pxl.cl/2VwqZ

The new config is added behind a ccaConfig and is not enabled by default.

Reviewed By: mjoras

Differential Revision: D47310050

fbshipit-source-id: 4ed1ef30f9734facdda237399af2250fabba071a