Add TransportSettingsFunctions for populating ccaConfig from JSON
Summary:
Adds a new header file that has functions to populate the CongestionControllerConfig struct from JSON.

The JSON format is a straightforward conversion that uses the same field names as the struct.

This will only be included by the server to parse incoming JSON knobs that are added in a separate change.

Reviewed By: hanidamlaj, mjoras

Differential Revision: D47309440

fbshipit-source-id: b56f48e101227146644a0c4bf60efab6b3a3880f