Purge old Rx timestamps based on ACKs of ACKs
Summary:
Prune stored Rx timestamps of packets if those packets were ACKed previously (`WriteAckFrame`) and we received an ACK of that ACK from the peer. `commonAckVisitorForAckFrame` already purges these packet numbers from AckBlocks in AckState. If a packet doesn't exist in the AckState, the timestamp shouldn't be sent as it was confirmed to be received.

TODO: this isn't true for old packets that are pruned in the AckState so some old timestamps might be pruned, which is probably OK!

This is being done to avoid sending duplicate timestamps and reduce network and memory overhead.

Reviewed By: bschlinker

Differential Revision: D49216621

fbshipit-source-id: a5b94ba974edac224c111cd82d7b7099312b52ac