Improve accounting of extraAckedDelivered in BBRv2
Summary:
Previously extraAckedDelivered was not reset for every train of packets so it could keep increasing and pushing the congestion window higher while in startup.

This did not have an impact on the speed the pacekts are sent because pacing was still only controlled by the bandwidth estimate_ but this could be visibile in the divergence between the cwnd and target pacing rate.

Reviewed By: hanidamlaj

Differential Revision: D49286012

fbshipit-source-id: 1340c939fde6b35e3edfefd248b52396c4c3195f