ensure CMAKE_INSTALL_LIBDIR is initialized
Summary:
`CMAKE_INSTALL_LIBDIR` is the canonical variable name for specifying the name of the directory where 'library' files go to, but it's defined by `GNUInstallDirs` which we currently don't include: https://cmake.org/cmake/help/latest/module/GNUInstallDirs.html

This was not detected in the Fedora build as it sets the override when calling cmake, but OSS builds using getdeps.sh fails.

Fix by setting this to a default if undefined.

Reviewed By: davide125

Differential Revision: D49400795

fbshipit-source-id: e30af569ce7b02c0a3f484c1eb8e0650f47f104a