Fix QuicClientTransport::recvmsg on Windows
Summary:
`QuicClientTransport::recvmsg` sets `msg.msg_name` to `nullptr` if we already have the remote address, but leaves `msg.msg_namelen` set to `kAddrLen`.

Linux seems to be ok with this combination — I assume without inspecting kernel sources that it ignores the length if the pointer is null. Windows is unhappy and breaks. Given the current way the code is written, the first packet gets received on Windows and subsequent packets fail.

I caught this while working on standardizing the receive path in D48715162; UTs caught this issue after I made it the default path. We should have had parameterized tests exploring all the paths before, but at least the simplification will resolve.

Reviewed By: mjoras, kvtsoy

Differential Revision: D48869381

fbshipit-source-id: 334875c6f3a974262a1778870caa96b0a39273fb