Add cca config options to ignore inflightHi and to control response to loss
Summary:
inflightHi currently forces the cwnd to be very close to the target BDP even during ProbeBW, this can slow down the ProbeBW_UP attempting to probe for more bandwidth.

Reaction to packet loss is a core difference between BBRv1 and BBRv2.

This adds a couple of cca config options to enable experimenting with both and compare specific parts of BBRv2 behavior in comparison to BBRv1.

Reviewed By: mjoras

Differential Revision: D49526118

fbshipit-source-id: 35ba83e834b1febd79c0e46af39b0205cde4dc1c