Adjust bandwidth windowed filter length to compensate for its use as expiry time
Summary:
The window_length parameter in the WindowedFilter is enforced by checking the distance between new sample and the old one is greater than the window length.

For BBRv2 max bandwidth filter, the timestamp is the integer cycle count number. Effectively the max value becomes the maximum of window+1 cycles.

This change compensates for that logic by subtracting 1 from the window length.

Reviewed By: mjoras

Differential Revision: D50136527

fbshipit-source-id: fd8a7bc444895e486f0a707c7aa2888ae1213083