add support for IP_BIND_ADDRESS_NO_PORT in AsyncSocketTransportFactory and SocketUtil
Summary:
IP_BIND_ADDRESS_NO_PORT socket option can be set only with SOL_IP level but it can be applied to both ipv4 and ipv6 sockets. Currently in both places, it assumes incorrectly that socket options with SOL_IP can only be applied to ipv4 socket so IP_BIND_ADDRESS_NO_PORT was not applied to ipv6 socket as expected.
The issue for TCP was fixed with the diff D19425729 and this diff is fixing the same for UDP.
For now, IP_BIND_ADDRESS_NO_PORT is the only socket option we are aware of which uses SOL_IP for both ipv4 and ipv6 sockets. In case we find other socket options with the same issue later, we will add handling later.

Reviewed By: jbeshay

Differential Revision: D51710738

fbshipit-source-id: 49906eb04bc43c93548c12324bafdbd2c0acb1ec