Introduce StreamIdSet and use it for open stream tracking.
Summary:
We "lazily" allocates streams. For example, suppose that we receive a stream frame for stream 400 before receiving any lower stream. This implicitly opens all streams of that type below 400. Instead of allocating a stream state for each, we track that those streams are open and then only allocate the stream state once the stream is actually used.

However, the way we store the open stream IDs is wasteful. It is stored as an individual entry in an F14FastSet. So if there are 400 open streams, we allocate 400 * 8 = 3200 bytes of space.

We can exploit the fact that these streams are effectively contiguous integers in the set [0, 4, 8, ... 400], storing them instead in an IntervalSet containing [0, 400]

(Note: this ignores all push blocking failures!)

Reviewed By: sharmafb

Differential Revision: D52259837

fbshipit-source-id: e777215fe4958e6526190df714b96008fdcb1a3a