Combine HighResQuicTimer and TimerFDQuicTimer into one type
Summary: Breaking these two types apart in D51413481 and using them separately for the client and server resulted in the TimerFD not being used anymore for non-mobile clients even when it's available. The combines them back into one type and restore the compile-time selection to use the TimerFD when not on mobile.

Reviewed By: sharmafb

Differential Revision: D52570625

fbshipit-source-id: 3f6117d67558f4506f9fe9efd9d0ce0ad2efbbbc