Modularity for packet group writing
Summary:
In this commit, I'm creating a `PacketGroupWriter`, which is an abstract class that has APIs for packet writing. I'm also creating a `UdpSocketPacketGroupWriter` which inherits from it.

The abstract `PacketGroupWriter` enables us to use different mechanisms to write packets, and is a prelude to us using AF_XDP for packet writing.

Note that this diff doesn't change any existing functionality.

Reviewed By: mjoras

Differential Revision: D52978211

fbshipit-source-id: 1d24f984f819d25f91c6fd762a5a40003b4c86ce