Round the batch size to the nearest integer in experimental mode
Summary:
The previous batch size calculation in the experimental pacer takes the floor of the batch size after scaling it up to the time since last write. Given the batchSizes can be as small as 5, taking the floor can significatly impact the pacer's ability to fully utilize the link.

This changes the batch size calculation to round the number to the nearest integer, which improves the accuracy of the pacer.

When trying to pace at 284 Mbps on a 500 Mbps link:
- Vanilla pacer achieves 128 Mbps
- Experimental pacer (without this change) achieves 273.7 Mbps
- Experimental pacer (with this change) achieves 282.5 Mbps

Reviewed By: mjoras

Differential Revision: D53549757

fbshipit-source-id: 2c81c48bd862acdf83a58c0ca1d993520adc4388