Improve probe bw cycle tracking
Summary:
Updating the cycle count in adaptUpperBounds() mean that it could happen prematurely if ProbeBw_Down lasts longer than a single round (rtt round).

This changes the logic to advance the ProbeBW cycle only when entering ProbeBw_Down. It also removes the advanceCycleAfterStartup option which is now irrelevant.

This results in better coexistence with loss-based CCAs.

Reviewed By: mjoras

Differential Revision: D53586407

fbshipit-source-id: fdc7ac23184e4edebb74f10026acb821f9df34f9