Have only pickXsk API in XskContainer
Summary:
Previously, the user would call functions like `getXskBuffer` and `writeXskBuffer` on the `XskContainer`. Now, I'm changing it so that we call `pickXsk`, and then call the relevant functions on the `XskSender`.

The reason I'm making this change is that it makes things cleaner with versions of the `XskContainer` that don't use hashing, since they don't need the source and destination addresses as input for all functions.

Reviewed By: hanidamlaj

Differential Revision: D53536905

fbshipit-source-id: 026fcdcb94b3f60bb905467998c0b3ce11880cdd