Client key update: Add an option to initiate the first key update a fixed number of packets into the connection [1/x]
Summary:
In order to exercise the key update path more frequently, this adds an option to initiate one key update early into the connection after a fixed number of packets is written. After that, keys are updated periodically at a significantly lower frequency.

Currently the defaults are to initiate the first key update after 500 packets, then periodically every 1M packets (which is under the confidentiality limit).

Reviewed By: mjoras

Differential Revision: D54082404

fbshipit-source-id: 3edd7489ceb8c643a319edb06e006803f1e736ba