Improve LibevQuicEventBase handling for runInLoop and thisIteration
Summary:
Previously, runInLoop would just call the callback function on top of the stack which was not the expected behavior in some of the call sites.

This change borrows the logic for handling runInLoop correctly from folly::EventBase.

runInEventBaseThreadAndWait() is not used in mvfst-mobile so this changes it to explicitly fail rather than calling the callback function immediately.

This change also adds a new set of unit tests to confirm that this behavior is consistent between LibevQuicEventBase and FollyQuicEventBase.

Reviewed By: kvtsoy

Differential Revision: D54646914

fbshipit-source-id: cc735f82d2f9a5c786a4abe65caa235bbd60149e