Do not delay work via runInLoop to the next loop iteration
Summary:
So mjoras, jbeshay and myself were looking at some weird rtt spikes and ack delayes when running a large download tests with mnscurl against a localhost hq server.
The ack delayes and rtt spikes look like this:
 {F1467932366}
Basically, we were frequently not sending acks in time from the client, and that resulted in choppy bytes flow.

Appears that we are using a slightly incorrect callback/watcher in our event base impl - we are using `ev_check` which apparently runs only inthe next loop iteration, and that is contingent on another event arriving and waking up the loop:
 {F1467951626}

What this means in practice we could be running into situations where we received a large batch of data, processed it and are ready to write our some acks, but with `ev_check` this write will be postponed to the next read event, which may not arrive soon given we already received good amount of data.

Replacing it with `ev_prepare` should do the trick, because this one runs right before the current loop goes to sleep, and our writes would not be contingent on another reads arriving anymore:
 {F1467952163}

http://pod.tst.eu/http://cvs.schmorp.de/libev/ev.pod#code_ev_prepare_code_and_code_ev_che

So with this change, the ack delays are gone:
 {F1467952739}

Reviewed By: jbeshay, mjoras

Differential Revision: D54819182

fbshipit-source-id: 043a9f9ecadf4fa382ddfa8f2e7e8d07e5d64efa