Fix initialization for largestNonDsrAcked to be a sequence number rather than a packet number
Summary:
This solves a rare corner case where packet reordering happens and an ack doesn't have any new acked packets. In this case, largestNonDsrAcked ends up with a packet number instead of a sequence number. Later on this produces very high reorder distance values and declares any out of order packets as lost (effectively threshold=1).

We need to initialize largestNonDsrAcked to a valid sequence number when there are no acked packets.

If we don’t initialize it at all, we can’t detect packet losses that occur across acks, i.e. ACK for non-dsr packet 1, followed by ACK for non-dsr packet 4 won’t trigger a loss by reorder.

If we initialize it to the largestAck packet num, we end up marking any out of order packet as a loss by reorder because the gap is always big.

To fix this properly, we need to track largestAcked dsr and non-dsr separately in the ackState rather than rely on translating from the packet number to the virtual sequence numbers inside the loss detection functions when no new packets are acked.

This change adds a new variable to track the largest acked non dsr sequence number and uses that to initialize largestNonDsrAcked.

Reviewed By: mjoras

Differential Revision: D55271688

fbshipit-source-id: fe59f9412745ae992fdd680d9d11039f67895b8c