Refactor ack visitor into separate packet and frame visitors
Summary:
The current ackVisitor passed to processAckFrame() gets executed with every acked frame.

Not every check in that visitor needs to be performed for every frame.

This refactors it into two different visitors, one that is called once per acked packet and another that is called once per acked frame.

Reviewed By: mjoras

Differential Revision: D56090734

fbshipit-source-id: 7b0877684e439f9e88c0aae7a294245570cf8309