Early return in stateless reset parsing
Summary:
As in title. If we parse something as a stateless reset matching our token, return early. The code before would check if the header parsed as a short header, but that's not correct. If the token bytes match, we should treat it as a stateless reset as this is a more certain check.

See: https://www.rfc-editor.org/rfc/rfc9000.html#section-10.3.1

Reviewed By: sharmafb

Differential Revision: D56376777

fbshipit-source-id: 2a7ce041c872f296db6c4d608895c3b81e479628