Use back pressure from congestion controller for mvfst mobile egress
Summary:
There is currently a problem with uploads that we're seeing in mvfst mobile QEs - basically mvfst transactions get cut short after 120s and we get a receive timeout.

That's because we set the read timer (120s) here https://fburl.com/code/818fwy3t after MNS egress queue is emptied.
With ngtcp2, it does already check the CC state before taking the bytes off the queue, but with mvfst we simply consume all the bytes from HTTP/MNS layer and store those in transport while egressing.

What happens is the timer above gets triggered right away for mvfst mobile, and we basically have 120s to both egress the body for the upload (which can take long time in highly congested networks) and get the response back.
For ngtcp2 that timer is triggered much later - pretty much once we egressed most of the body - which gives ngtcp2 much longer to finish the transaction and to not fall victim of the "read timeout".

The fix is to introduce a similar back pressure driven by congestion controller to mvfst mobile write path.

Reviewed By: mjoras

Differential Revision: D56791402

fbshipit-source-id: 3e18122e99afca451bc131dce280ae4ab78fcc20