Track function loop callbacks to ensure they are deleted with LibevQuicEventBase
Summary:
runInLoop(Func) works by creating a FunctionLoopCallback that wraps the function into a QuicEventBaseLoopCallback.

For a normal use of QuicEventBaseLoopCallback, the object is owned by the class that creates it and it manages its lifetime.

The special FunctionLoopCallbacks are owned by the LibevQuicEventBase. They are destroyed when the callback occurs. Previously, if the Eventbase was destroyed before calling all these callbacks, these wrappers leaked. This change ensures they are destroyed with the EventBase.

Reviewed By: kvtsoy

Differential Revision: D56860623

fbshipit-source-id: be57452cde5f3454bf20512beee389dfc468e307