Personalized CWND Initial Setup
Summary: Set up the foundation for personalized cwnd. During the `onReadData` method, the `cwndHint` from the client will be checked. At that point the client will be classified as a "weak", "moderate", or "strong" connectivity client. Assuming there is at least one sample RTT and the new cwnd value is larger than the current one, the congestion controller will be reinitialized with the higher cwnd value.

Reviewed By: jbeshay, pfarcasanu

Differential Revision: D57211407

fbshipit-source-id: 2157d02cf0baf1d7b31d617cf4801584d8e1c678