Ensure psk is removed from the client cache when the early data paramters don't match the ones from the handshake
Summary:
If early data is rejected and the tls parameters that we used from the 0-rtt ticket do not match the ones in the current handshake, the connection cannot continue.

Previously, the connection would fail but the psk still stayed in the client cache. This meant that subsequent retries could still attempt zero rtt and hit the same error repeatedly.

This change moves the decision to close the connection to the QuicClientTransport instead of the FizzClientHandshake so it can remove the "bad" psk from the pskCache.

Reviewed By: kvtsoy

Differential Revision: D57398519

fbshipit-source-id: b1c076794b9b16954ec23474bffc5a0be7e11090