Add RX delay metric
Summary:
We do not necessarily read all received packets in a single event base loop iteration, so the latency of the loop itself may be misleadingly low, if packets are queued up on the socket. To add visibility into this state, add an "RX delay" metric, the amount of time it takes between NIC receiving the packet (timestamped using
SO_TIMESTAMPING socket option) and time packet was received from the socket in onDataAvailable().

Reviewed By: jbeshay, mjoras

Differential Revision: D57808455

fbshipit-source-id: 19db8f2e1c2af038892d13568b96444eb9347be8