Convert Ack Rx timestamps disable list to allow list
Summary:
There are still some timestamps seen on dynamic content/connections and disabling timestamps per hostname doesn't scale well.
Instead, this diff will disable timestamps by default in Liger and enable only for static content connections based on a hostname filter (`scontent`, `video`, `cdninstagram`). The actual meat of the change is in `AdvancedHTTPSessionManager::shouldEnableQuicAckRxTimestamps` - everything else is mostly a replacement of disable to enable.

This is implemented for MNS in D58099124.

Reviewed By: jbeshay

Differential Revision: D58099125

fbshipit-source-id: 06118163fcc5d2e2ce90028810d6a9af5c7958a9