Add an API to FollyQuicEventBase to run a function immediately if the API is called in the EventBase thread and not wait if it's not called there
Summary: We have `QuicEventBase::runImmediatelyOrRunInEventBaseThreadAndWait` which runs a function immediately if it is called in the `EventBase` thread.  However, if it is not called in the `EventBase` thread the `FollyQuicEventBase::runImmediatelyOrRunInEventBaseThreadAndWait` implementation runs it in the right `EventBase` but waits for its completion, as the API name denotes.  Please note that `LibevQuicEventBase` does not actually implement this API.  We have use cases where we want to run a function immediately, if we are in the right `EventBase` but don't want to wait if we are not.  `folly::EventBase` has such an API and this patch adds one, `QuicEventBase::runImmediatelyOrRunInEventBaseThread`, to use it.  `LibevQuicEventBase` won't have a real implementation as is the case with its implementation of `QuicEventBase::runImmediatelyOrRunInEventBaseThreadAndWait`.

Reviewed By: jbeshay

Differential Revision: D58315330

fbshipit-source-id: 101e117f9c0e9e8ef117e3073c814811a162a3ca