add intermediary copy to avoid strict aliasing shenanigans
Summary:
This introduces extra copies to avoid strict aliasing violations and avoid potential compiler optimization headaches.

Specifically, in the previous code `storage` was aliasing `key`, which makes the operation of `getAddress` which does the `memcpy` liable to be altered.

By copying the values explicitly to local variables it is not valid for any optimization passes to mess with them.

An alternative that would also probably work is making a `getAddressRaw` which does the memcpy through an `unsigned char*` instead since the compiler cannot assume a lack of aliasing for those pointer types.

Reviewed By: hanidamlaj

Differential Revision: D58822149

fbshipit-source-id: 1e28c3cb6e20d20622347d4f6233e6fc2664f653